Rename handlers.go to userHandlers.go for better abstraction and readability
